We created the root container and added a the class for it.
Then we started by generating the HTML for the tag as a string and returns an HTML element for it. We use that to map all tags to elements using the map function on the tags array. Then we use that to map all tags to elements using the map function on the tags array.

Then we mapped the element to HTML by using the outerHTML attribute from it.

With the container element created and the HTML for the tags ready, the innerHTML attribute of the content we set using a template string and returned it.