We are creating a list of products and displaying them on our food product web page. Each product is represented by an object with properties such as name, price, unit, tags, and image. The code consists of several functions that work together to create the list of products.
The createContent function takes a product object as an argument and creates a div element with the class name content. The innerHTML of this element is set to a template literal that includes the product’s name, price, unit, and tags. The tags are created by mapping over the tags property of the product object, calling the createTagElement function on each tag, and then joining the resulting array of HTML strings into a single string. The function returns the content element

The createProductItem function takes a product object as an argument and creates a div element with the class name item. This element is then appended with the result of calling the createProductImage function with the image property of the product object as an argument, and the result of calling the createContent function with the product object as an argument. The function returns the item element.

The createProductImage function takes an image source string as an argument and creates a div element with the class name image. An img element is then created and its src attribute is set to the image source string. The image is then appended to the image container and the image container is returned.

The createTagElement function takes a tag string as an argument and creates a div element with the class name 'ui label teal'. The innerText of this element is set to the tag string and the element is returned.

The createListForProducts function takes an array of product objects as an argument. It selects an element with the class name 'items', sets its innerHTML to an empty string, and then appends an item element for each product in the array by calling the createProductItem function on each product.

Finally, at the end of the code, the createListForProducts function is called with an array of products as an argument to create and display the list of products on the web page.